Module jaffi_support::exceptions

Structs

AnyThrowable

Error

Exception

A type that represents a known Exception type from Java.

Traits

Throwable

Functions

catch_panic_and_throw

Catches and potential panics, and then converts them to a RuntimeException in Java.

get_panic_message

register_panic_hook

This panic hook can add a bit more information than the catch_unwind, which doesn’t get the full panic_info